WilsonServer: Moved to /impl and defined interface
To keep with the interface, impl. design - I've declared
an interface for WilsonServer and created an implementation
of it in /impl.
