+++ src/wilson/interfaces/IWilsonServer.ts
@@ -1,5 +1,10 @@
 import ServerEvent from "../types/ServerEvent";
 
+declare interface IWilsonServer {
+    on(event: "undeliverable", listener: (event: ServerEvent) => void): this;
+    emit(event: "undeliverable", payload: ServerEvent): boolean;
+}
+
 interface IWilsonServer {
     sendEvent(event: ServerEvent, clientUuid: string): void;
 }
